+++ src/main/java/sic/modelo/Pedido.java
@@ -3,6 +3,7 @@ package sic.modelo;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import lombok.Data;
@@ -24,7 +25,7 @@ public class Pedido implements Serializable {
     private String nombreUsuario;
     private List<Factura> facturas;        
     private List<RenglonPedido> renglones;
-    private double totalEstimado;
-    private double totalActual;
+    private BigDecimal totalEstimado;
+    private BigDecimal totalActual;
     private EstadoPedido estado;
 }
