+++ src/main/java/sic/modelo/Proveedor.java
@@ -1,6 +1,7 @@
 package sic.modelo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -23,7 +24,7 @@ public class Proveedor implements Serializable {
     private Localidad localidad;
     private Empresa empresa;
     private boolean eliminado;
-    private Double saldoCuentaCorriente; 
+    private BigDecimal saldoCuentaCorriente; 
     private Date fechaUltimoMovimiento;
 
     @Override
